unify node config overwrite
